5 acceptance testing

+++ app/controllers/questions_controller.rb
@@ -1,19 +1,23 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   def index
     @questions = Question.all
   end
 
-  def show; end
+  def show
+    @answer = question.answers.new
+  end
 
   def new; end
 
   def create
-    @question = Question.create(question_params)
+    @question = Question.create(question_params.merge(user_id: current_user.id))
 
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question was successfully created.'
     else
       render :new
     end
@@ -30,8 +34,13 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question successfully deleted!'
+    else
+      flash[:error] = 'You are not allowed delete the answer'
+      redirect_to questions_path
+    end
   end
 
   private
